remove authorization checks for auth0 & renamed function/file name to match what it does
